Bug 1373202 — Moved BrowserProfileSyncDelegate into AppDelegate.
… and rename it to AppSyncDelegate.

This is done because we want to start using more functionality from the AppDelegate.

This touches all profiles initializer calls.
